requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-74hc595==1.4.5
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-charlcd==3.4.12
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-framebuf==1.6.7
adafruit-circuitpython-mcp230xx==2.5.15
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-ssd1306==2.12.19
adafruit-circuitpython-typing==1.11.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
annotated-types==0.7.0
anyio==4.6.2.post1
binho-host-adapter==0.1.6
board==1.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorzero==2.0
elevenlabs==1.12.1
exceptiongroup==1.2.2
gpiozero==2.0.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
lgpio==0.2.2.0
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.2
pigpio==1.78
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyftdi==0.56.0
pygame==2.6.1
pyserial==3.5
pytesseract==0.3.13
python-dotenv==1.0.1
pyusb==1.2.1
requests==2.32.3
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPLCD==1.3.1
simpleaudio==1.0.4
smbus2==0.5.0
sniffio==1.3.1
spidev==3.6
sysv-ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1